Clean up test, provide a temp. solution to the "out of tsuck" issue
It appears that the "out of stack" issue is called by one of the rules.
My suspicion is that the rules is "fold/augment", which is used by (++).
For the moment I used a custom (++) function, which makes things work it seems.

darcs-hash:20080309031901-5035a-2a2ecfd349e50566640a7d5aef3e8db4602888d0.gz